+++ internal/completion/database.go
@@ -2,6 +2,7 @@ package completion
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/service"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/request"
@@ -42,6 +43,34 @@ func (s DatabaseType) CompleteArgument(ctx context.Context, svc service.AllServi
 	var vals []string
 	for _, t := range dbTypes {
 		vals = append(vals, t.Name)
+	}
+	return MatchStringPrefix(vals, toComplete, true), cobra.ShellCompDirectiveNoFileComp
+}
+
+// DatabaseProperty implements argument completion for database properties.
+type DatabaseProperty struct {
+	ServiceType string
+}
+
+// make sure DatabaseType implements the interface
+var _ Provider = DatabaseProperty{}
+
+// CompleteArgument implements completion.Provider
+func (s DatabaseProperty) CompleteArgument(ctx context.Context, svc service.AllServices, toComplete string) ([]string, cobra.ShellCompDirective) {
+	dbType, err := svc.GetManagedDatabaseServiceType(ctx, &request.GetManagedDatabaseServiceTypeRequest{Type: s.ServiceType})
+	if err != nil {
+		return None(toComplete)
+	}
+
+	properties := dbType.Properties
+	var vals []string
+	for key, details := range properties {
+		description := details.Title
+		if description == key && details.Description != "" {
+			description = details.Description
+		}
+
+		vals = append(vals, fmt.Sprintf("%s\t%s", key, description))
 	}
 	return MatchStringPrefix(vals, toComplete, true), cobra.ShellCompDirectiveNoFileComp
 }